requirements.txt
# Audio capture and processing
soundcard>=0.4.2
pyaudio>=0.2.11
numpy>=1.24.0
scipy>=1.10.0

# Speech recognition and VAD
openai-whisper>=20230918
webrtcvad>=2.0.10
pydub>=0.25.1

# System tray and GUI
pystray>=0.19.4
Pillow>=10.0.0

# HTTP requests
requests>=2.31.0
aiohttp>=3.8.5

# Threading and async
asyncio-mqtt>=0.16.1

# Cross-platform support
pywin32>=306; sys_platform == 'win32'
pyobjc-framework-Cocoa>=9.0; sys_platform == 'darwin'

# Utilities
python-dotenv>=1.0.0
psutil>=5.9.5
